Add more explicit dependency of modules
During destroy we noticed IGW being deleted before the EKS cluster is
brought down which causes IGW to not proceed until EKS is deleted. This is
an attempt to fix that by giving explicit module dependencies.

This caused the `apply` to begin a little late. Also add a note that the
root module should not be used for prod management as is.

Also remove AWS Provider config from `client-vpn` module

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>