Move away from data token to file based for kubernetes and helm auth
It seems the token is cached by terraform and when it expires terraform
does not fetch new one; hence all helm/kubernetes provider operation fails
Switching to the kubeconfig generated by `eks` module saves us that trouble

This made the terraform code more reliable; but at the same time, it
just introduced the trouble of maintaining the state in the kubeconfig
file as part of apply and destroy.

Also make VPN creation optional.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>