Remove explicit inter-dependency of modules
The stack seems to be consistent without the explicit dependency. Removed
it since as per hashicorp/terraform#27523 this is not necessary.

This introduces inconsistencies in destroy process, make sure to keep
running it repeatedly and ignore the kube config error, since destroying
EKS will destroy the helm resources in any case.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>