Upgrade all provider and module versions to latest ones
The most notable upgrade is kubernetes provider and EKS module upgrades.
The kubernetes provider upgrade needed the authentication process to be
updated by moving from file based kube config to AWS CLI based token
generation. This was needed because after upgrading seemingly tf plan needs
file to be present during plan and hence the switch. The EKS module needed
default volume type to be specified and we defaulted it to `gp2`.

Signed-off-by: Imran M Yousuf <imyousuf@gmail.com>